A new approach to FibonacciNumbers is here.

+++ src/test/java/assignments/ReplitAnswers/FibonacciNumbers.java
@@ -13,7 +13,7 @@ public class FibonacciNumbers {
 			
 		  }
 		  
-		  public static int fib(int num){
+		  public static long fib(int num){
 		    //WRITE YOUR CODE HERE
 			  if (num <= 1) {
 		            return num;
